ctrl-c test: remove auxiliary file, use new process group
