+++ Lib/test/test_asyncio/test_windows_events.py
@@ -39,19 +39,60 @@ class UpperProto(asyncio.Protocol):
 
 class ProactorLoopCtrlC(test_utils.TestCase):
     def test_ctrl_c(self):
-        from .test_ctrl_c_in_proactor_loop_helper import __file__ as f
+        code = """if 1:
+        import asyncio
+        import ctypes
+        import sys
+        from ctypes import wintypes
+        
+        # child process is created with CREATE_NEW_PROCESS_GROUP flag
+        # an implicit call to SetConsoleCtrlHandler(NULL,TRUE) will be made
+        # on behalf of new process which will disable CTRL-C in it
+        # call SetConsoleCtrlHandler(NULL, False) to restore normal processing
+        SetConsoleCtrlHandler = ctypes.windll.kernel32.SetConsoleCtrlHandler
+        SetConsoleCtrlHandler.argtypes = (ctypes.POINTER(ctypes.c_voidp), wintypes.BOOL)
+        SetConsoleCtrlHandler.restype = wintypes.BOOL
+        SetConsoleCtrlHandler(None, False)
 
-        # ctrl-c will be sent to all processes that share the same console
-        # in order to isolate the effect of raising ctrl-c we'll create
-        # a process with a new console
-        flags = subprocess.CREATE_NEW_CONSOLE
-        with spawn_python(f, creationflags=flags) as p:
+        asyncio.set_event_loop_policy(asyncio.WindowsProactorEventLoopPolicy())
+        l = asyncio.get_event_loop()
+
+        def step(n):
             try:
-                exit_code = p.wait(timeout=5)
-                self.assertEqual(exit_code, 255)
+                print(n)
+                sys.stdout.flush()
+                l.run_forever()
+                sys.exit(100)
+            except KeyboardInterrupt:
+                # ok
+                pass
             except:
+                sys.exit(200)
+
+        step(1)
+        step(2)
+        sys.exit(255)
+        """
+        def step(p, expected):
+            s = p.stdout.readline()
+            self.assertEqual(s, expected)
+            # ensure that child process gets to run_forever
+            time.sleep(0.5)
+            os.kill(p.pid, signal.CTRL_C_EVENT)
+        
+        exit_code = None
+        flags = subprocess.CREATE_NEW_PROCESS_GROUP
+        with spawn_python("-c", code, creationflags=flags) as p:
+            try:
+                step(p, b"1\r\n")
+                step(p, b"2\r\n")
+                exit_code = p.wait(timeout=5)
+            except Exception as e:
                 p.kill()
-                raise
+                p.wait()
+                stdout, stderr = p.communicate()
+                self.fail(f"{stdout: stdout}\nstderr:{stderr}")
+        self.assertEqual(exit_code, 255)
 
 
 class ProactorTests(test_utils.TestCase):
